création d'un article

+++ src/main/java/com/eni/amis/des/objets/bll/AdresseService.java
@@ -7,7 +7,7 @@ import com.eni.amis.des.objets.bo.Adresse;
 public interface AdresseService {
 
     Adresse create(Adresse adresse);
-    Adresse findById(long id);
+    Adresse findById(int id);
     void update(Adresse adresse);
     List<Adresse> getAllAddresses();
