+++ src/main/java/com/eni/amis/des/objets/dal/AddressDAO.java
@@ -7,7 +7,7 @@ import com.eni.amis.des.objets.bo.Adresse;
 public interface AddressDAO {
 
     Adresse create(Adresse adresse);
-    Adresse findById(long id);
+    Adresse findById(int id);
     void update(Adresse adresse);
     List<Adresse> findAll();
